% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunctions.R
\name{do_car}
\alias{do_car}
\title{Use three inputs to predict response using R.}
\usage{
do_car(r, rm, m1, m2, c1, c2, date)
}
\arguments{
\item{r}{individual stock return(numeric vector)}

\item{rm}{market return(numeric vector)}

\item{m1}{days of the start of the model estimation period before the event(numeric)}

\item{m2}{days of the end of the model estimation period before the event(numeric)}

\item{c1}{days of the start of the model window period before the event(numeric)}

\item{c2}{days of the start of the model window period before the event(numeric)}

\item{date}{date of the event(date)}
}
\value{
a random sample of size \code{n}
}
\description{
Calculate the cumulative abnormal return of the stocks using event study method.
}
\examples{
\dontrun{
x<-'XF'
y<-'HS300'
d1<-'2019-01-02'
d2<-'2020-12-12'
getSymbols("x",from=d1,to=d2)
getSymbols("y",from=d1,to=d2)
r <- periodReturn(X$`X.Adjusted`,period='daily',type='log')
rm <- periodReturn(Y$`Y.Adjusted`,period='daily',type='log')
do_car(r,rm,10,5,1,1,'2020-01-02')
}
}
