% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{rwMetropolis using Rcpp}
\usage{
rwMetropolis(sigma, x0, n)
}
\arguments{
\item{sigma}{sigma of normal distribution}

\item{x0}{initial value}

\item{n}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
rwMetropolis using Rcpp
}
\examples{
\dontrun{
rwMetropolis(1,50,10)
}
}
